Go 1.22, jemalloc 5.3.0 and go.mod updates
Upgrades go version to 1.22, jemalloc to 5.30, and performed a go get -u all on go.mod dependencies.

Note: github.com/uber/jaeger-client-go could only be upgraded to 2.28.0+incompatible (not 2.30) as the interface has changed.

Fixes all of the many critical and high vulnerabilities in the image arising from go (not OS image)